requirements.txt
opencv-python-headless
numpy
torch
ultralytics 
pillow
streamlit 
transformers